# AD14: Restructure loop to fix optimization bug